Fix get_512_bit_chunks, add another test case

+++ src/util.py
@@ -16,7 +16,7 @@ def get_next_multiple_of_512(num: int) -> int:
 def get_512_bit_chunks(binary:str) -> typing.List[str]:
     if (len(binary) % 512 != 0):
         raise ValueError('size of binary should be a multiple of 512')
-    total_chunks = len(binary) / 512
+    total_chunks = int(len(binary) / 512)
     chunks = [''] * total_chunks
     for i in range(total_chunks):
         chunks[i] = binary[(i*512):((i+1)*512)]
